Fixed filtering system, updated ai system
Fixed collaborative filtering's query timeout issue, and modified ai chatbot system to: only recalculate vectors on server start, add an option so that it just loads vectors from .pkl file, meaning it doesn't have to recalculate anything (very fast startup), and added content filtering for the games it recommends. The .pkl's are not uploaded to github, so new server hosters will have to run the vectorization calculations on first startup.